requirements.txt
# requirements.txt - Final Cleaned Version
streamlit
databricks-sqlalchemy==2.0.8
SQLAlchemy>=2.0.41,<3
numpy==1.26.4
protobuf==4.25.3
langchain
langchain-core
langchain-community
langchain-openai
python-dotenv
altair